refactor: extract 5 sub-view renderers, thin LastSimRecapView to orchestrator (1.29 Phase 3)
